create match service and calculate match final score fix functions that had wrong calculation and include king and queen information on each player
